Fix image detection in client not decoding original URL
The worker is decoding proxy URLs to use the original URL since we want to process the original images, not already processed images.
It then uses the original URL as the key to save the signed proxy URLs.

The client however does not do this. So for proxy URLs, this check always fails since the worker never uses proxy URLs as the key.

This commit fixes that during populating the image cache and when querying it.