Fix local prod builds assuming CDN (#814)
* Fix local prod builds assuming CDN

Prod builds assumed that we're running in an AWS environment and use a CDN. This commit changes that.

Now, if the AWS way to fetch the commit failed _and_ the normal git command fails and only if, we assume we're running the prod build locally and don't configure the CDN.

* Fix path to app_version_manifest.json

Was autoformatted by linter. Probably before I added eslint-next-disable-line above.