Enhanced the cursor to be stateful
Fixes #147

To avoid duplicated results push some state towards the cursor.

The cursor uses a Bloom Filter[1], a probabilistic data structure. In this data
structure, False positive matches are possible, but false negatives are not –
in other words, a query returns either "possibly in set" or "definitely not in
set". Elements can be added to the set, but not removed. The more items added,
the larger the probability of false positives.

[1] https://en.wikipedia.org/wiki/Bloom_filter