Remove unused onPaid for RECEIVE paid action (#2229)
This function can be removed because it will never do anything when called.

It does not do anything for wrapped invoices, and it will never get called for direct payments (since we don't know when the invoice was paid).

These are the only two payment methods for the RECEIVE paid action since 5a8804d.