PoC for authorization with schema directives
This is only a proof of concept how to guard fields with schema directives so it is only applied to UserPrivates.

It can be tested using this GraphQL query:

```
fragment UserFields on User {
  id
  name
  privates {
    sats
  }
}

query users {
  me {
    ...UserFields
  }
  user(id: 624) {
    ...UserFields
  }
}
```

However, I am not sure how schema directives can support conditional validation like we need for fields that are conditionally private.
Conditional validation might need to continue exist as custom code in the resolvers.

Next steps could be to also apply these directives to limit API key usage.